Divided services between mongo and influx
